+++ badageul/badageul/Property List.plist
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>UILaunchStoryboardName</key>
+	<string>Launch Screen</string>
+	<key>Fonts provided by application</key>
+	<array>
+		<string>Fonts/GowunDodum-Regular.ttf</string>
+	</array>
+</dict>
+</plist>
